Extend full admin privileges on sudo
Introduce the notion of sudo to enable admins to behave like normal users, but opt in to using their full privileges.

This prototype introduces the `/plugins/playbooks/api/v0/sudo` endpoint. A `POST` here when logged in as an administrator creates a KV entry keyed by the session id and marked to expire in 30 minutes. Whenever we check permissions find the user to be an administrator, we then check for the presence of this sudo bit to determine whether to keep those admin privileges, or downgrade to a normal user. At no point does the sudo bit ever grant a user more access, even if somehow applied to a non-admin.

Fixes: https://mattermost.atlassian.net/browse/MM-35659